+++ packages/transaction-pay-controller/src/strategy/relay/relay-quotes.test.ts
@@ -1,5 +1,6 @@
 import { successfulFetch } from '@metamask/controller-utils';
 import type { TransactionMeta } from '@metamask/transaction-controller';
+import type { Hex } from '@metamask/utils';
 import { cloneDeep } from 'lodash';
 
 import {
@@ -12,7 +13,10 @@ import { getRelayQuotes } from './relay-quotes';
 import type { RelayQuote } from './types';
 import { NATIVE_TOKEN_ADDRESS } from '../../constants';
 import { getMessengerMock } from '../../tests/messenger-mock';
-import type { QuoteRequest } from '../../types';
+import type {
+  GetDelegationTransactionCallback,
+  QuoteRequest,
+} from '../../types';
 import { calculateGasCost, calculateTransactionGasCost } from '../../utils/gas';
 import { getNativeToken, getTokenFiatRate } from '../../utils/token';
 
@@ -25,14 +29,14 @@ jest.mock('@metamask/controller-utils', () => ({
 }));
 
 const QUOTE_REQUEST_MOCK: QuoteRequest = {
-  from: '0x123',
+  from: '0x1234567890123456789012345678901234567891',
   sourceBalanceRaw: '10000000000000000000',
   sourceChainId: '0x1',
   sourceTokenAddress: '0xabc',
   sourceTokenAmount: '1000000000000000000',
   targetAmountMinimum: '123',
   targetChainId: '0x2',
-  targetTokenAddress: '0xdef',
+  targetTokenAddress: '0x1234567890123456789012345678901234567890',
 };
 
 const QUOTE_MOCK = {
@@ -62,12 +66,12 @@ const QUOTE_MOCK = {
           },
           data: {
             chainId: 1,
-            data: '0x123',
-            from: '0x1',
+            data: '0x123' as Hex,
+            from: '0x1' as Hex,
             gas: '21000',
             maxFeePerGas: '1000000000',
             maxPriorityFeePerGas: '2000000000',
-            to: '0x2',
+            to: '0x2' as Hex,
             value: '300000',
           },
           status: 'complete',
@@ -78,7 +82,20 @@ const QUOTE_MOCK = {
   ],
 } as RelayQuote;
 
-const TRANSACTION_META_MOCK = {} as TransactionMeta;
+const TRANSACTION_META_MOCK = { txParams: {} } as TransactionMeta;
+
+const DELEGATION_RESULT_MOCK = {
+  authorizationList: [
+    {
+      chainId: '0x1' as Hex,
+      nonce: '0x2' as Hex,
+      yParity: '0x1' as Hex,
+    },
+  ],
+  data: '0x111' as Hex,
+  to: '0x222' as Hex,
+  value: '0x333' as Hex,
+} as Awaited<ReturnType<GetDelegationTransactionCallback>>;
 
 describe('Relay Quotes Utils', () => {
   const successfulFetchMock = jest.mocked(successfulFetch);
@@ -90,8 +107,11 @@ describe('Relay Quotes Utils', () => {
     calculateTransactionGasCost,
   );
 
-  const { messenger, getRemoteFeatureFlagControllerStateMock } =
-    getMessengerMock();
+  const {
+    messenger,
+    getDelegationTransactionMock,
+    getRemoteFeatureFlagControllerStateMock,
+  } = getMessengerMock();
 
   beforeEach(() => {
     jest.resetAllMocks();
@@ -115,6 +135,8 @@ describe('Relay Quotes Utils', () => {
       cacheTimestamp: 0,
       remoteFeatureFlags: {},
     });
+
+    getDelegationTransactionMock.mockResolvedValue(DELEGATION_RESULT_MOCK);
   });
 
   describe('getRelayQuotes', () => {
@@ -160,6 +182,52 @@ describe('Relay Quotes Utils', () => {
             tradeType: 'EXPECTED_OUTPUT',
             user: QUOTE_REQUEST_MOCK.from,
           }),
+        }),
+      );
+    });
+
+    it('includes transactions in request', async () => {
+      successfulFetchMock.mockResolvedValue({
+        json: async () => QUOTE_MOCK,
+      } as never);
+
+      await getRelayQuotes({
+        messenger,
+        requests: [QUOTE_REQUEST_MOCK],
+        transaction: {
+          ...TRANSACTION_META_MOCK,
+          txParams: {
+            data: '0xabc' as Hex,
+          },
+        } as TransactionMeta,
+      });
+
+      const body = JSON.parse(
+        successfulFetchMock.mock.calls[0][1]?.body as string,
+      );
+
+      expect(body).toStrictEqual(
+        expect.objectContaining({
+          authorizationList: [
+            {
+              chainId: 1,
+              nonce: 2,
+              yParity: 1,
+            },
+          ],
+          tradeType: 'EXACT_OUTPUT',
+          txs: [
+            {
+              to: QUOTE_REQUEST_MOCK.targetTokenAddress,
+              data: '0xa9059cbb0000000000000000000000001234567890123456789012345678901234567891000000000000000000000000000000000000000000000000000000000000007b',
+              value: '0x0',
+            },
+            {
+              to: DELEGATION_RESULT_MOCK.to,
+              data: DELEGATION_RESULT_MOCK.data,
+              value: DELEGATION_RESULT_MOCK.value,
+            },
+          ],
         }),
       );
     });
@@ -325,8 +393,8 @@ describe('Relay Quotes Utils', () => {
       });
 
       expect(result[0].fees.targetNetwork).toStrictEqual({
-        usd: '1.23',
-        fiat: '2.34',
+        usd: '0',
+        fiat: '0',
       });
     });