+++ packages/transaction-pay-controller/src/types.ts
@@ -17,6 +17,7 @@ import type { NetworkControllerFindNetworkClientIdByChainIdAction } from '@metam
 import type { NetworkControllerGetNetworkClientByIdAction } from '@metamask/network-controller';
 import type { RemoteFeatureFlagControllerGetStateAction } from '@metamask/remote-feature-flag-controller';
 import type {
+  AuthorizationList,
   TransactionControllerAddTransactionBatchAction,
   TransactionControllerUnapprovedTransactionAddedEvent,
 } from '@metamask/transaction-controller';
@@ -59,6 +60,11 @@ export type TransactionPayControllerGetStateAction = ControllerGetStateAction<
   TransactionPayControllerState
 >;
 
+export type TransactionPayControllerGetDelegationTransactionAction = {
+  type: `${typeof CONTROLLER_NAME}:getDelegationTransaction`;
+  handler: GetDelegationTransactionCallback;
+};
+
 /** Action to get the pay strategy type used for a transaction. */
 export type TransactionPayControllerGetStrategyAction = {
   type: `${typeof CONTROLLER_NAME}:getStrategy`;
@@ -78,6 +84,7 @@ export type TransactionPayControllerStateChangeEvent =
   >;
 
 export type TransactionPayControllerActions =
+  | TransactionPayControllerGetDelegationTransactionAction
   | TransactionPayControllerGetStateAction
   | TransactionPayControllerGetStrategyAction
   | TransactionPayControllerUpdatePaymentTokenAction;
@@ -93,6 +100,9 @@ export type TransactionPayControllerMessenger = Messenger<
 
 /** Options for the TransactionPayController. */
 export type TransactionPayControllerOptions = {
+  /** Callback to convert a transaction into a redeem delegation. */
+  getDelegationTransaction: GetDelegationTransactionCallback;
+
   /** Callback to select the PayStrategy for a transaction. */
   getStrategy?: (
     transaction: TransactionMeta,
@@ -399,3 +409,15 @@ export type UpdatePaymentTokenRequest = {
   /** Chain ID of the new payment token. */
   chainId: Hex;
 };
+
+/** Callback to convert a transaction to a redeem delegation. */
+export type GetDelegationTransactionCallback = ({
+  transaction,
+}: {
+  transaction: TransactionMeta;
+}) => Promise<{
+  authorizationList?: AuthorizationList;
+  data: Hex;
+  to: Hex;
+  value: Hex;
+}>;